Création du ClientService Tests unitaires de l'entité Client et de ClientService

+++ src/main/java/fr/eni/caveavin/bo/Client.java
@@ -13,11 +13,12 @@ import lombok.*;
 @Entity
 @Table(name = "CAV_CLIENTS")
 public class Client {
+    // TODO contraintes validation
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "ID")
-    private Long id;
+    private int id;
 
     @EqualsAndHashCode.Include
     @Column(name = "LOGIN", nullable = false, unique = true, length = 100)
